requirements.txt
monai>=1.3.0
openslide-python>=1.3.1
opencv-python>=4.8.0
numpy>=1.23.0
pillow>=10.0.0
pandas>=2.0.0
pyyaml>=6.0.1
matplotlib>=3.7.0
streamlit>=1.33.0
scikit-learn>=1.3.0
torchcam>=0.4.0
